+++ frontend/src/screens/Blockchain/StakingKey/Pool/Tabs/index.js
@@ -1,11 +1,10 @@
 import React from 'react'
 import {defineMessages} from 'react-intl'
-import {makeStyles} from '@material-ui/styles'
 import {Grid} from '@material-ui/core'
 import {LiteTabs, LiteTab, LoadingInProgress} from '@/components/visual'
 import {useI18n} from '@/i18n/helpers'
-import HistoryTab from '../../common/History'
-import TransactionsTab from '../../common/TransactionsTab'
+import History from '../../common/History'
+import Transactions from '../../common/Transactions'
 import {LoadingError, TabsPaginationLayout} from '@/components/common'
 import {useLoadStakepoolHistory} from '../dataLoaders'
 
@@ -19,24 +18,25 @@ export const TAB_NAMES = {
   TRANSACTIONS: 'TRANSACTIONS',
 }
 
+const _History = ({stakepool}) => {
+  const {error, loading, data: stakepoolHistory} = useLoadStakepoolHistory(stakepool.hash)
+  return error ? (
+    <LoadingError error={error} />
+  ) : loading ? (
+    <LoadingInProgress />
+  ) : (
+    <History history={stakepoolHistory} />
+  )
+}
+
+const _Transactions = ({stakepool}) => <Transactions transactions={stakepool.transactions} />
+
 const TABS_CONTENT = {
-  [TAB_NAMES.HISTORY]: ({stakepool}) => {
-    const {error, loading, data: stakepoolHistory} = useLoadStakepoolHistory(stakepool.hash)
-    return error ? (
-      <LoadingError error={error} />
-    ) : loading ? (
-      <LoadingInProgress />
-    ) : (
-      <HistoryTab history={stakepoolHistory} />
-    )
-  },
-  [TAB_NAMES.TRANSACTIONS]: ({stakepool}) => (
-    <TransactionsTab transactions={stakepool.transactions} />
-  ),
+  [TAB_NAMES.HISTORY]: _History,
+  [TAB_NAMES.TRANSACTIONS]: _Transactions,
 }
 
 const StakepoolTabs = ({stakepool, pagination, tabState}) => {
-  // const classes = useStyles()
   const {translate: tr} = useI18n()
   const {setTabByEventIndex, currentTabIndex, currentTab} = tabState
   const TabContent = TABS_CONTENT[currentTab]
